refactor: renamed files and changed Register page
