+++ frontend/src/app/register/page.tsx
@@ -1,97 +1,9 @@
 "use client";
 
-import { useState } from "react";
-import { CountryCombobox } from "@/features/auth/components/CountryCombobox";
-import {
-  Field,
-  FieldLabel,
-  FieldGroup,
-  FieldSeparator,
-  FieldSet,
-} from "@/shared/components/ui/field";
-import { Input } from "@/shared/components/ui/input";
-import { Button } from "@/shared/components/ui/button";
-import {
-  type CountryCallingCode,
-  getCountryCallingCode,
-  parsePhoneNumberWithError,
-  ParseError,
-} from "libphonenumber-js";
-import { Phone } from "lucide-react";
+import { useRegisterStep } from "@/features/auth/hooks/useRegisterStep";
+import RegisterForm from "@/features/auth/components/RegisterForm";
 
 export default function Register() {
-  const [phoneNumber, setPhoneNumber] = useState("");
-  const [callingCode, setCallingCode] = useState<CountryCallingCode>(
-    getCountryCallingCode("PL")
-  );
-  const [error, setError] = useState<Error | null>(null);
-
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    try {
-      const phoneNumberObject = parsePhoneNumberWithError(
-        `+${callingCode}${phoneNumber}`
-      );
-      if (!phoneNumberObject.isValid()) {
-        throw new Error("Enter valid phone number!");
-      }
-
-      console.log(phoneNumberObject);
-      // Make a request to the api with tanstack query passing phone number
-    } catch (err) {
-      if (err instanceof ParseError) {
-        if (err.message.includes("TOO_SHORT")) {
-          setError(new Error("Phone number is too short."));
-        } else if (err.message.includes("TOO_LONG")) {
-          setError(new Error("Phone number is too long."));
-        } else {
-          setError(new Error("Enter a valid phone number!"));
-        }
-      } else {
-        setError(new Error("An unknown error occurred."));
-      }
-    }
-  };
-
-  return (
-    <div className="h-screen flex justify-center items-center">
-      <form className="w-full max-w-xs md:max-w-md" onSubmit={handleSubmit}>
-        <FieldGroup>
-          <FieldSet>
-            <h1 className="text-3xl">Register</h1>
-            <FieldSeparator />
-            <FieldGroup>
-              <Field>
-                <FieldLabel>Country</FieldLabel>
-                <CountryCombobox
-                  setCallingCode={setCallingCode}
-                ></CountryCombobox>
-              </Field>
-              <Field>
-                <FieldLabel>Phone Number</FieldLabel>
-                <div className="flex rounded-md border border-input focus-within:ring-2 focus-within:ring-offset-0 focus-within:ring-primary">
-                  <div className="flex items-center px-3 text-sm bg-muted text-muted-foreground rounded-l-md border-input">
-                    <Phone className="mr-2 h-4 w-4 opacity-70" />+{callingCode}
-                  </div>
-                  <Input
-                    type="tel"
-                    value={phoneNumber}
-                    onChange={(e) => {
-                      setError(null);
-                      setPhoneNumber(e.target.value);
-                    }}
-                    className="rounded-l-none focus-visible:ring-0 border-none"
-                  />
-                </div>
-                {error && <h3 className="text-destructive">{error.message}</h3>}
-              </Field>
-            </FieldGroup>
-          </FieldSet>
-          <Field>
-            <Button>Register</Button>
-          </Field>
-        </FieldGroup>
-      </form>
-    </div>
-  );
+  const { step } = useRegisterStep();
+  return step === "register" ? <RegisterForm /> : <div>Verify step</div>;
 }
