+++ frontend/src/features/auth/lib/sendPhoneNumber.ts
@@ -1,5 +1,7 @@
 import { fetchWrapper } from "@/lib/fetchWrapper";
+import { buildUrl } from "@/lib/buildUrl";
 
 export const sendPhoneNumber = async (phoneNumber: string) => {
-  return fetchWrapper("/api/auth/register/initiate", "POST", { phoneNumber });
+  const url = buildUrl("/auth/register/initiate");
+  return fetchWrapper(url, "POST", { phoneNumber });
 };
